app/requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.112.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.7
magic-filter==1.0.12
motor==3.5.1
multidict==6.0.5
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pymongo==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.8
setuptools==72.1.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.17.6
watchgod==0.8.2
websockets==12.0
yarl==1.9.4
